Add support for configurable PIN prompting.
Add support for checking to see if we should set the flag
PROTECTED_AUTHENTICATION_PATH in the token information field,
based on user preferences.

This preference is found in our application domain "mil.navy.nrl.cmf.pkcs11"
and uses the key "askPIN".  It can either be a string (for a single
application name) or an array (for multiple applications).