Use approriate access control operation
We were cheating a bit before and we were always using the
same access control operation when evaluating the PIN, specifically
LAAccessControlOperationUseKeySign.  It turns out that was wrong; for
keys where signing is not permitted, that will return an error.  Not
quite sure how this worked before; I must have really messed something
up!  Use the appropriate access control operation for each key.