Change to explicit alloc/init call for password
Change the creation of NSData object that holds the password string
to explicitly calling the alloc/init methods, rather than calling
the convenience method dataWithBytes.

I didn't realize that under Cocoa naming conventions, because the
convenience function didn't start with "alloc" or "copy", I didn't
technically own that object and it was added to the autorelease
pool.  This would be fine if we were writing everything in Objective-C,
but the Crypoki is a C API and I can't guarantee people to call the
pool release methods at some point.  It turns out most applications
never did and it worked fine, but when Firefox would do it I'd get
a crash because of the release message I had originally sent to the
password object.

I think the most robust solution is to stick with explicit allocate/release
for this object.