Use -shrext to change module suffix
Normally libtool will build shared libraries on MacOS X with a
suffix of ".dylib", but if you tell it to build a module with -module
it will make the suffix default to .so.  Because a lot of programs
want to link a shared library with .dylib on MacOS X, I had the Makefile
create a symlink ending in .dylib to the .so file.  But I then discovered
the -shrext option to libtool that let you set the shared library
extension directly, so I switched to that.