Fix application identifier for dylib signing
Sigh.  Somehow I wasn't paying attention and instead of using the
APPIDENTIFIER as the identifier prefix for signing the dylib, I used
APP_SIGNING_ID, which is the hash of the signing certificate.  Which ...
is unique, I guess, so it's not wrong, but I meant to use the hash
identifier.  Fix it for the NEXT release.