Move sleep to top of loop
We were running into problems during notarization where getting the
request UUID and then immediately requesting the notarization status
would occasionally result in a error reponse which would say, "invalid
UUID".  To combat this, move the "sleep 10" to the top of the loop so
we start out by sleeping 10 seconds, THEN checking for the notarization
status.  Notarization usually takes a minute or two, so this shouldn't
change how long we have to wait.